+++ godot/3d-quests/scripts/get_image.gd.uid
@@ -1,0 +1,1 @@
+uid://bjwux5glfx0y1
