+++ src/app/maps/page.tsx
@@ -1,0 +1,9 @@
+import AddMap from "@/components/ui/add-map";
+export default function AddMapPage() {
+
+    return (
+        <div>
+            <AddMap />
+        </div>
+    );
+}
